requirements.txt
ddash==0.3.0
gunicorn==23.0.0
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
